Refactor/improvement of the SelectLink component
